making the follower row click into profile
