+++ frontend/src/LoggedIn/FollowersList/followerPanel.css
@@ -5,13 +5,18 @@
   position: fixed;
   right: 0;
   z-index: 150;
-  padding: 2.4rem 2rem 4rem;
   overflow: hidden;
   display: flex;
   flex-direction: column;
 }
 
-.test {
+.following-title {
+  display: flex;
+  justify-content: space-between;
+  padding: 2.4rem 2rem 1rem;
+}
+
+.text-ellipsis {
   text-overflow: ellipsis;
   white-space: nowrap;
   overflow: hidden;
@@ -29,4 +34,26 @@
 
 .searchModal {
   height: 400px;
+}
+
+.follow-row {
+  color: white;
+  text-decoration: none;
+  padding: 0.25rem 2rem;
+}
+.follow-row:hover {
+  background-color: rgba(255, 255, 255, 0.7);
+  color: black;
+}
+
+.follow-btn {
+  background-color: #FFCC86;
+  border-color:#FFCC86;
+  border-radius: 6px;
+  border-width: 1px;
+  padding: 0.5rem 2rem;
+}
+.follow-btn:hover {
+  background-color: #CFA161;
+  border-color: #CFA161;
 }
